Terminal client prototype
For now just processing args and parsing reference. Syntax is

slash.bash [-f file] reference[:type][:docset]

My intention is to store a config file with at least url to the
slash server. The optional -f flag is to override default config
file.

We can search for references providing optionally type and/or docset.
For now we won't manage to specify more types or docsets in a single
query.

My intention is to directly visualize the result when only one result
is returned, or a list of results when several are. This way we maintain
the terminal client API simpler and there's no ambiguity with links
included in the docs, since providing all three values reference, type
and docset, a maximum of one result can be returned.