Bash: menu to select result when multiple results
The problem with this is that testing became very hard due to
difficulties to mock calls to dialog.

I implemented multiple mocking abilities in xunit but finally had
to comment out previous tests that are now broken although the code
is working properly. Must revisit xunit mocking code to fix mocking.