Docset interface refactored.
There was a little misconception about promise and lazy loading. In knex
there is the same misconception and that guided us to make the same
design errors.It should be more clear now. You build the query and
execute it with method execute. Execute returns you a promise.

That's all!