Reimplement /api/get by uri
Now uri is the identity used to get the results.

Since routes for web client now have to match anything (to capture
multiple slash-separated components) I had to reorder the routes
and add specific routes for the static 'css' and 'js' resources.

I also fixed the tests and the sample data in the database because
we were handling some cases with # in the uris, but express ignores
whatever goes after the #, so for instance "test.html" and "test.html#test_2"
are the same uri for express, just to mention some of the several
test cases we had.