TreeView & fix uris
Implemented TreeView component. For now is not very sofisticated:
it's prepared to show any hierarchy of nodes containing 'name' and
'children' (being these also nodes), but for now I'm just querying
for the children of the very first parent of the current resource,
that is, current resource siblings. I think it could be very
interesting to add in the future also all the children from the
current node, which will allow us to browse the whole docset from
the TreeView.

To query the children I made a /api/children method. Following
same convention than existing methods, the url should contain the
uri of the resource whose children we're interested in. In the
backend two DB queries are needed to implement this: first getting
the id corresponding to that uri, and then getting all references
with parent_id equal to that id. Alternatively we could make a single
query with a join. I tried both approaches and had difficulties.
I couldn't honestly implement the join in a simpler manner within
this fluent api we made for the docsets. On the other hand, I
didn't find the way to return a single promise built from two
secquencial promises in the new slash.children() method, so I
finally splitted this into two methods (get_id and children)
and built the secquence in the router.

I fixed the sample data for the tests and the tests themselves,
as well as some of the backend and the frontend code, dealing
with the uris. The problem was that the uris we have stored in
the database have changed their format. Now they start with a
slash (/) and they have %20 instead of spaces. This format is
currently the best we found to allow compatibility with markdown
links, and probably with something else in the future. Some
code we had (both in tests and implementation) dealt with uris
as if they had spaces and not an initial slash, so I made a bunch
of fixes. A core part of the problem is that even when we use
%20 instead of spaces in a http query, it magically translates
them to spaces, so somewhere we have to make the translation
the other way round. I put the fix in the slash methods "get" and
"get_id" where I transform the "identity" adding the initial slash
and replacing spaces with %20. This should probably end up being
a decode_url method which could be in another place between the
router and the access to the database.