Client-side JADE templates
To compile them it's necessary to install clientjade, which is only
available to install globally, so you should install it in your local
machines:

  sudo npm install -g clientjade

All JADE templates under views/ are processed to build the client-side
templates to public/js/jade-templates.js. This is done automatically
when running grunt (before the task 'develop'), although I wasn't be
able to make it work with the 'watch' mode, so for now it's necessary
to rerun Grunt each time we change a JADE template for the client side.

I extracted to JADE templates all rendering logic of TreeView and
OutlineView, so these two entities are now much simpler. I also
added some useful code to Reference and made subtle fixes in the
linking of parents with childrens.

Now the designer will be very happy of avoiding any interaction with
code :-)