Added spider process to parser job.
First we made a refactor. Now we have 4 components:

Spider -> Parser -> Styler -> Docsets (DAO)

The spider walks the web from given url and pass to parser a map
of uris and theirs htmls. The parser processes each of html and return
a object wiith references and uris replacements. The Styler picks this
collections and apply some changes to content. At the end the Docsets
DAO persist all the references in the database.

Bufff, a lot of work.