Reference.get('objects')
With this .get('objects').each(function(object) { ... }) syntax we
allow asynchronous work on client side, because retrieved objects
can really be a lot.

For now I didn't change the JADE templates and what I do is simply
re-render the views when a new object is added to the reference data.

TreeView and OutlineView are now a little redundant and should
probably be merged in a single component.