Fix bug - pagination without order
According to postgres guide plan executions can changed
from time to time if you don't set a order different to default.
http://www.postgresql.org/docs/current/static/queries-limit.html

With this change the results order returned by search was very ugly.
I improved a little the search algorithm. But later we have to apply
some of this patterns:
http://blog.2ndquadrant.com/text-search-strategies-in-postgresql/