Fix errors with setting sync
Why:

The way it is implemented is very error-prone, in addition it needs a
big refactor. But first of all fix load incorrect settings and prevent
to corrupt the cookie

This change addresses the need by:

Change settings store where sync is made, and remove wrong logic in
settings store infrastructure.